embed-fastapi/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
open-clip-torch==2.24.0
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy==1.26.3
prometheus-client==0.19.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
httpx==0.26.0
aiofiles==23.2.1
